# graph-api/api.yaml
openapi: 3.0.3
info:
  title: Graph API
  version: "1.0"
tags:
  - name: ui
    description: Для использования юзером
  - name: service
    description: Для программного взаимодействия
paths:
  /is-unique:
    get:
      tags: [service]
      operationId: isUnique
      parameters:
        - in: query
          name: url
          schema:
            type: string
            format: uri
          required: true
      responses:
        200:
          description: true if unique, false otherwise
          content:
            application/json:
              schema:
                type: boolean
  /find-path:
    get:
      tags: [ui]
      operationId: findPath
      parameters:
        - in: query
          name: src
          required: true
          schema:
            type: string
            format: uri
        - in: query
          name: dst
          required: true
          schema:
            type: string
            format: uri
      responses:
        200:
          description: path found
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uri
              example:
                - "https://ru.wikipedia.org/wiki/%D0%97%D0%B0%D0%B3%D0%BB%D0%B0%D0%B2%D0%BD%D0%B0%D1%8F_%D1%81%D1%82%D1%80%D0%B0%D0%BD%D0%B8%D1%86%D0%B0"
                - "https://ru.wikipedia.org/wiki/%D0%97%D0%B0%D0%B3%D0%BB%D0%B0%D0%B2%D0%BD%D0%B0%D1%8F_%D1%81%D1%82%D1%80%D0%B0%D0%BD%D0%B8%D1%86%D0%B0"
                - "https://ru.wikipedia.org/wiki/%D0%97%D0%B0%D0%B3%D0%BB%D0%B0%D0%B2%D0%BD%D0%B0%D1%8F_%D1%81%D1%82%D1%80%D0%B0%D0%BD%D0%B8%D1%86%D0%B0"
        400:
          description: Src or dst not found
        404:
          description: Path from src to dst not found
  /count:
    get:
      tags: [ui, service]
      operationId: getCount
      responses:
        200:
          description: Количество урлов в графе
          content:
            application/json:
              schema:
                type: integer
                format: int32